# openapi.yaml
openapi: 3.1.0
info:
  title: TaskLane API
  version: 0.1.0
  description: Single source of truth API specification for TaskLane Mini Kanban Board.
paths:
  /api/tasks:
    get:
      summary: List tasks with optional filtering and search
      operationId: listTasks
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskStatus'
          description: Filter tasks by Kanban column status
        - name: priority
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskPriority'
          description: Filter tasks by priority level
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive search across task title and description
      responses:
        '200':
          description: A list of tasks sorted newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation error for query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create a new task (strictly placed in backlog)
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully in backlog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation error (e.g. empty title or client-supplied status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

  /api/tasks/{task_id}:
    get:
      summary: Get task details by ID
      operationId: getTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Partially update a task (e.g. column status transition or priority)
      operationId: updateTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error in request body or path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete a task by ID
      operationId: deleteTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid path parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'

components:
  schemas:
    TaskStatus:
      type: string
      enum:
        - backlog
        - todo
        - in_progress
        - review
        - done
      description: The 5 discrete sequential workflow states

    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high
        - urgent
      description: Task priority rating

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          description: Concise task summary (cannot be empty or whitespace-only)
        description:
          type: string
          default: ""
          description: Optional detailed notes
        priority:
          $ref: '#/components/schemas/TaskPriority'
          default: medium
      additionalProperties: false

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
      additionalProperties: false

    TaskResponse:
      type: object
      required:
        - id
        - title
        - description
        - status
        - priority
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string

    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object